refactor : 이전 레벨에서 사용하던 dto 모두 제거
