+++ src/main/java/chess/dto/response/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 public class ErrorResponseDto {
