+++ src/main/java/chess/service/ChessService.java
@@ -10,7 +10,7 @@ import chess.domain.game.Turn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.dto.BoardsDto;
-import chess.dto.StatusDto;
+import chess.dto.response.StatusResponseDto;
 import chess.dto.request.MoveRequestDto;
 import chess.dto.request.RoomRequestDto;
 import chess.dto.response.GameResponseDto;
@@ -110,9 +110,9 @@ public class ChessService {
         roomRepository.updateGameOver(id);
     }
 
-    public StatusDto createStatus(final Long id) {
+    public StatusResponseDto createStatus(final Long id) {
         final Board board = toBoard(boardRepository.findBoardByRoomId(id));
-        return StatusDto.of(new Score(board.getBoard()));
+        return StatusResponseDto.of(new Score(board.getBoard()));
     }
 
     private void validateGameOver(final RoomEntity room) {
